requirements.txt
numpy
torch
pyzmq